+++ app/task/moph-appointment.js
@@ -74,22 +74,30 @@ async function markAsSent(row) {
         console.error('Error marking as sent:', error);
     }
 }
-const process = async () => {
+const process = async (date = null) => {
     await createAppointmentTable();
     await cleanOldRecords();
     hospitalConfig = await (0, moph_refer_1.getHospitalConfig)();
-    if (!hospitalConfig || !hospitalConfig.configure || !hospitalConfig.configure?.moph_appointment || hospitalConfig.configure?.moph_appointment?.enable != 1) {
-        console.error(moment().format('HH:mm:ss'), 'MOPH Appointment Process: Appointment Service Disabled');
+    let isAlertAppointment = false;
+    if (hospitalConfig && hospitalConfig?.configure && hospitalConfig.configure?.moph_appointment) {
+        isAlertAppointment = hospitalConfig.configure?.moph_appointment?.alert_after_service == 1 || hospitalConfig.configure?.moph_appointment?.alert_before == 1 ? true : false;
+    }
+    if (!isAlertAppointment) {
+        console.error(moment().format('HH:mm:ss'), 'MOPH Alert for Appointment Process: Appointment Service Disabled');
         return false;
     }
-    const date = moment().subtract(30, 'minutes').format('YYYY-MM-DD 00:00:00');
-    return await getData(date);
+    date = date || moment().subtract(30, 'minutes').format('YYYY-MM-DD');
+    console.log('');
+    const result = await getData(date);
+    console.log('-'.repeat(70));
+    return result;
 };
 async function getData(date) {
     try {
         date = moment(date).format('YYYY-MM-DD');
         let opdVisit = await hismodel_1.default.getAppointment(db, 'visit_date', date);
         if (opdVisit.length > 0) {
+            console.log(moment().format('HH:mm:ss'), 'MOPH Appointment Process Date:', date, 'Found Records:', opdVisit.length);
             let skippedCount = 0;
             let sentResult = [];
             for (let row of opdVisit) {
@@ -125,7 +133,8 @@ async function getData(date) {
         }
     }
     catch (error) {
-        throw error;
+        console.error(moment().format('HH:mm:ss'), 'MOPH Appointment Process Error:', error.message || error);
+        return error;
     }
 }
 exports.default = { process };